# contracts/java-openapi.yaml
openapi: 3.1.0
info:
  title: GreenBook Community API - Agent Facade
  version: 1.0.0
  description: |
    GreenBook 社区后端 Agent Facade 稳定 API。
    所有接口使用 RS256 JWT Bearer Token 鉴权。

    ## 鉴权
    - JWKS 公钥端点: `/.well-known/jwks.json` (无需认证)
    - Token 必须包含 audience: `greenbook-agent-runtime`
    - token_type 必须为 `access`
    - Python 配置: `GREENBOOK_JAVA_BASE_URL=http://127.0.0.1:8080`
  contact:
    name: GreenBook Team
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT Access Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: 幂等键，用于写操作的幂等保证
      required: false
      schema:
        type: string
        example: "uuid-or-unique-key"
...
  schemas:
...
    AgentPostContext:
      type: object
      properties:
        postId:
          type: string
        title:
          type: string
        description:
          type: string
        body:
          type: string
          description: 正文内容 (Markdown)
        tags:
          type: array
          items:
            type: string
        authorId:
          type: string
        authorNickname:
          type: string
        publishTime:
          type: string
          format: date-time
        contentOrigin:
          type: string
          enum: [MANUAL, AI_ASSISTED]

    # ── My Posts ──────────────────────────────────────────
...
    AgentDraftCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 256
        content:
          type: string
          description: 正文 Markdown
        summary:
          type: string
          maxLength: 200
        visibility:
          type: string
          enum: [public, followers, school, private, unlisted]
...
    DraftResponse:
      type: object
      properties:
        draftId:
          type: string
        ownerId:
          type: string
        title:
          type: string
        content:
          type: string
        summary:
          type: string
        tags:
          type: array
          items:
            type: string
        visibility:
          type: string
        version:
          type: integer
        status:
          type: string
        contentOrigin:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Scheduled Publication ─────────────────────────────
...
paths:
...
  /api/v1/agent/posts/{postId}:
    get:
      tags: [Agent Facade]
      summary: 获取帖子详情
      operationId: getPost
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 帖子详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentPostContext'

  # ── My Posts ──────────────────────────────────────────
...
  /api/v1/agent/drafts:
    post:
      tags: [Agent Facade]
      summary: 创建草稿
      operationId: createDraft
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentDraftCreateRequest'
      responses:
        '201':
          description: 草稿创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftResponse'
...